fix: move functions related to TemplateString to TSR-types.
This is to allow consumers of TSR-types to be able to work with TSR-types data (for example `url: string | TemplateString`) without having to import TSR.
