+++ packages/timeline-state-resolver/src/lib.ts
@@ -8,7 +8,6 @@ import {
 	ActionExecutionResultCode,
 	TimelineDatastoreReferences,
 	ActionExecutionResult,
-	TemplateString,
 } from 'timeline-state-resolver-types'
 import * as _ from 'underscore'
 import { PartialDeep } from 'type-fest'
@@ -281,6 +280,7 @@ export function fillStateFromDatastore(state: Timeline.TimelineState<TSRTimeline
 	return filledState
 }
 
+/** Create a Translatable message */
 export function t(key: string, args?: { [k: string]: any }): ITranslatableMessage {
 	return {
 		key,
@@ -307,26 +307,4 @@ export function actionNotFoundMessage(id: never): ActionExecutionResult<any> {
 
 export function cloneDeep<T>(input: T): T {
 	return klona(input)
-}
-
-/**
- * Interpolate a translation style string
- */
-export function interpolateTemplateString(key: string, args: { [key: string]: any } | undefined): string {
-	if (!args || typeof key !== 'string') {
-		return String(key)
-	}
-
-	let interpolated = String(key)
-	for (const placeholder of key.match(/[^{}]+(?=})/g) || []) {
-		const value = args[placeholder] || placeholder
-		interpolated = interpolated.replace(`{{${placeholder}}}`, value)
-	}
-
-	return interpolated
-}
-
-export function interpolateTemplateStringIfNeeded(str: string | TemplateString): string {
-	if (typeof str === 'string') return str
-	return interpolateTemplateString(str.key, str.args)
 }
